refactor: renamed CallsiteCommand -> RawCommandCallsiteRecord; CallsiteCommandTransform -> RawCommandCallsiteRecordTransform
