+++ src/compiler/test-file/formats/raw.js
@@ -5,13 +5,13 @@ import TestFile from '../../../api/structure/test-file';
 import Fixture from '../../../api/structure/fixture';
 import Test from '../../../api/structure/test';
 import createCommandFromObject from '../../../test-run/commands/from-object';
-import { CallsiteCommand } from '../../../utils/callsite-command';
+import { RawCommandCallsiteRecord } from '../../../utils/raw-command-callsite-record';
 
 export default class RawTestFileCompiler extends TestFileCompilerBase {
     static _createTestFn (commands) {
         return async t => {
             for (let i = 0; i < commands.length; i++) {
-                const callsite = commands[i]?.id ? new CallsiteCommand(commands[i].id, commands) : commands[i]?.callsite;
+                const callsite = commands[i]?.id ? new RawCommandCallsiteRecord(commands[i].id, commands) : commands[i]?.callsite;
 
                 try {
                     const command = createCommandFromObject(commands[i], t.testRun);
