+++ src/services/serialization/replicator/create-replicator.ts
@@ -13,7 +13,7 @@ import FunctionMarkerTransform from './transforms/function-marker-transform';
 import PromiseMarkerTransform from './transforms/promise-marker-transform';
 import ConfigureResponseEventOptionTransform from './transforms/configure-response-event-option-transform';
 import URLTransform from './transforms/url-transform';
-import CallsiteCommandTransform from './transforms/callsite-command-transform';
+import RawCommandCallsiteRecordTransform from './transforms/raw-command-callsite-record-transform';
 
 const DEFAULT_ERROR_TRANSFORM_TYPE = '[[Error]]';
 
@@ -49,6 +49,6 @@ export default function (): Replicator {
             new RequestHookEventDataTransform(),
             new RoleTransform(),
             new CallsiteRecordTransform(),
-            new CallsiteCommandTransform(),
+            new RawCommandCallsiteRecordTransform(),
         ]);
 }
