+++ src/utils/prerender-callsite.ts
@@ -1,7 +1,7 @@
 import { CallsiteRecord } from 'callsite-record';
 import renderCallsiteSync from './render-callsite-sync';
 import createStackFilter from '../errors/create-stack-filter';
-import { CallsiteCommand } from './callsite-command';
+import { RawCommandCallsiteRecord } from './raw-command-callsite-record';
 import getRenderers from './get-renderes';
 
 export interface RenderedCallsite {
@@ -11,7 +11,7 @@ export interface RenderedCallsite {
     noColor: string;
 }
 
-export default function prerenderCallsite (callsite: CallsiteRecord | CallsiteCommand): RenderedCallsite {
+export default function prerenderCallsite (callsite: CallsiteRecord | RawCommandCallsiteRecord): RenderedCallsite {
     const stackFilter = createStackFilter(Error.stackTraceLimit);
     const renderers   = getRenderers(callsite);
