+++ src/utils/render-callsite-sync.js
@@ -1,5 +1,5 @@
 import { renderers as renderersRecord } from 'callsite-record';
-import { renderers as renderersCommand } from './callsite-command';
+import { renderers as renderersCommand } from './raw-command-callsite-record';
 
 
 export default function renderCallsiteSync (callsite, options) {
